refactor(ui): create useFiltersHandler hook for FilterBar
