+++ airflow-core/src/airflow/ui/src/utils/useFiltersHandler.ts
@@ -1,0 +1,55 @@
+/*!
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { useCallback } from "react";
+import { useSearchParams } from "react-router-dom";
+
+import { useTableURLState } from "src/components/DataTable/useTableUrlState";
+import type { FilterConfig, FilterValue } from "src/components/FilterBar";
+
+export const useFiltersHandler = (filterConfigs: Array<FilterConfig>) => {
+  const [searchParams, setSearchParams] = useSearchParams();
+  const { setTableURLState, tableURLState } = useTableURLState();
+  const { pagination, sorting } = tableURLState;
+
+  const handleFiltersChange = useCallback(
+    (filters: Record<string, FilterValue>) => {
+      filterConfigs.forEach((config) => {
+        const value = filters[config.key];
+
+        if (value === null || value === undefined || value === "") {
+          searchParams.delete(config.key);
+        } else {
+          searchParams.set(config.key, String(value));
+        }
+      });
+
+      setTableURLState({
+        pagination: { ...pagination, pageIndex: 0 },
+        sorting,
+      });
+      setSearchParams(searchParams);
+    },
+    [filterConfigs, pagination, searchParams, setSearchParams, setTableURLState, sorting],
+  );
+
+  return {
+    handleFiltersChange,
+    searchParams,
+  };
+};
